Add a 'no media' option when selecting attachment in page blocks
When selecting an attachment to add to a site page block, there is a
"quick add" feature that let the user select multiple items at once.
When this "quick add" feature is used, only the item get attached
(media_id IS NULL is site_block_attachment)

But once the item has been attached, if the user click on "Open
attachment options", and then click on "Apply changes" in the sidebar
without changing anything, the primary media is automatically selected.

So, without changing anything the user actually changed something
without noticing.

Now consider a public item with a single media which is private.
In the first scenario (no media is selected), an anonymous user visiting
the page would see the item title but not the media.
But in the second scenario (a media is selected), an anonymous user
won't be able to see the item at all

And once we are in the second scenario, the only way to go back to the
first is to remove the attachement, and re-add it using the "quick add"
feature.

This patch intends to fix that by providing a "no media" option that is
selected by default (instead of the primary media) when no media was
selected.

Fixes #1708